requirements.txt
pandas>=1.3.0
numpy>=1.21.0
requests>=2
openai>=2.8.1
biopython>=1.86
tiktoken>=0.1.1
google-genai
